Move env and log utilities to lib
