+++ api/src/v1/index.ts
@@ -1,3 +1,1 @@
 export * from './v1.module';
-export * from './env';
-export * from './log';
